Moved basis functions to basis._functions submodul
