+++ openghg_inversions/basis/__init__.py
@@ -1,1 +1,2 @@
+from ._functions import bucketbasisfunction, quadtreebasisfunction
 from ._wrappers import basis_functions_wrapper
